Renaming logstructured to sqllog
